refactor(ui): consolidate schedule interval (#32082)
* refactor(ui): consolidate schedule interval

* fix(ui): preserve app install schedule navigation

* fix(ui): format scheduler form wiring test

* fix(ui): update test suite scheduler e2e selectors

* fix(ui): expose selected scheduler frequency state

* fix(ui): stabilize data quality e2e selectors

* refactor(ui): extract scheduler helpers and inline styles

* fix(ui): update scheduler e2e coverage

* fix(ui): reject descending cron minute ranges

* fix(ui): align Playwright suppression baseline
