+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/AddDataQualityTest/components/TestCaseFormV1.less
@@ -102,7 +102,7 @@
       }
     }
 
-    .schedule-interval-v1 {
+    .schedule-interval {
       .cron-expression-card {
         .ant-card-body {
           padding: @padding-xs;
