+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/AddDataQualityTest/components/TestCaseSchedulerSection.tsx
@@ -25,7 +25,7 @@ import { useTranslation } from 'react-i18next';
 import { DEFAULT_SCHEDULE_CRON_DAILY } from '../../../../constants/Schedular.constants';
 import { Transi18next } from '../../../../utils/i18next/LocalUtil';
 import { escapeESReservedCharacters } from '../../../../utils/StringUtils';
-import ScheduleIntervalV1 from '../../../Settings/Services/AddIngestion/Steps/ScheduleIntervalV1';
+import ScheduleInterval from '../../../Settings/Services/AddIngestion/Steps/ScheduleInterval';
 import { AddTestCaseList } from '../../AddTestCaseList/AddTestCaseList.component';
 import { AddTestCaseListChangePayload } from '../../AddTestCaseList/AddTestCaseList.interface';
 import { TestCaseSchedulerSectionProps } from './TestCaseFormV1.interface';
@@ -193,7 +193,7 @@ const TestCaseSchedulerSection: FC<TestCaseSchedulerSectionProps> = ({
 
         <FormField control={form.control} name="cron">
           {({ field }) => (
-            <ScheduleIntervalV1
+            <ScheduleInterval
               defaultSchedule={DEFAULT_SCHEDULE_CRON_DAILY}
               entity={t('label.test-case')}
               includePeriodOptions={schedulerOptions}
