+++ openmetadata-ui/src/main/resources/ui/src/components/Settings/Applications/AppSchedule/AppSchedule.component.tsx
@@ -14,7 +14,6 @@ import { Button, Col, Modal, Row, Space, Typography } from 'antd';
 import { isEmpty } from 'lodash';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { SCHEDULAR_OPTIONS } from '../../../../constants/Schedular.constants';
 import { useLimitStore } from '../../../../context/LimitsProvider/useLimitsStore';
 import {
   AppScheduleClass,
@@ -25,7 +24,6 @@ import { getIngestionPipelineByFqn } from '../../../../rest/ingestionPipelineAPI
 import { getCronDefaultValue } from '../../../../utils/CronExpressionUtils';
 import Loader from '../../../common/Loader/Loader';
 import ScheduleInterval from '../../Services/AddIngestion/Steps/ScheduleInterval';
-import { WorkflowExtraConfig } from '../../Services/AddIngestion/Steps/ScheduleInterval.interface';
 import applicationsClassBase from '../AppDetails/ApplicationsClassBase';
 import AppRunsHistory from '../AppRunsHistory/AppRunsHistory.component';
 import { AppRunsHistoryRef } from '../AppRunsHistory/AppRunsHistory.interface';
@@ -47,6 +45,8 @@ const AppSchedule = ({
   const [isPipelineDeployed, setIsPipelineDeployed] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
   const [isSaveLoading, setIsSaveLoading] = useState(false);
+  const [scheduleValue, setScheduleValue] = useState<string>();
+  const [isScheduleValid, setIsScheduleValid] = useState(true);
   const { config } = useLimitStore();
   const isAppDisabled = disabled || Boolean(appData.deleted);
 
@@ -113,9 +113,17 @@ const AppSchedule = ({
     setShowModal(false);
   };
 
-  const onDialogSave = async (data: WorkflowExtraConfig) => {
+  const onDialogOpen = () => {
+    setScheduleValue(
+      (appData.appSchedule as AppScheduleClass)?.cronExpression ?? undefined
+    );
+    setIsScheduleValid(true);
+    setShowModal(true);
+  };
+
+  const onDialogSave = async () => {
     setIsSaveLoading(true);
-    await onSave(data.cron ?? '');
+    await onSave(scheduleValue ?? '');
     setIsSaveLoading(false);
     setShowModal(false);
   };
@@ -168,29 +176,16 @@ const AppSchedule = ({
     t,
   ]);
 
-  const { initialOptions, initialData, defaultCron } = useMemo(() => {
+  const { initialOptions, defaultCron } = useMemo(() => {
     return {
       initialOptions: applicationsClassBase.getScheduleOptionsForApp(
         appData.name,
         appData.appType,
         pipelineSchedules
       ),
-      initialData: {
-        cron: (appData.appSchedule as AppScheduleClass)?.cronExpression,
-      },
       defaultCron: getCronDefaultValue(appData?.name ?? ''),
     };
-  }, [appData.name, appData.appType, appData.appSchedule, pipelineSchedules]);
-
-  const translatedSchedularOptions = useMemo(
-    () =>
-      SCHEDULAR_OPTIONS.map((option) => ({
-        ...option,
-        title: t(option.title),
-        description: t(option.description),
-      })),
-    [t]
-  );
+  }, [appData.name, appData.appType, pipelineSchedules]);
 
   useEffect(() => {
     fetchPipelineDetails();
@@ -252,7 +247,7 @@ const AppSchedule = ({
                   data-testid="edit-button"
                   disabled={isAppDisabled}
                   type="primary"
-                  onClick={() => setShowModal(true)}>
+                  onClick={onDialogOpen}>
                   {t('label.edit')}
                 </Button>
               )}
@@ -287,19 +282,28 @@ const AppSchedule = ({
         title={t('label.update-entity', { entity: t('label.schedule') })}
         width={650}>
         <ScheduleInterval
-          isEditMode
-          buttonProps={{
-            cancelText: t('label.cancel'),
-            okText: t('label.save'),
-          }}
           defaultSchedule={defaultCron}
           includePeriodOptions={initialOptions}
-          initialData={initialData}
-          schedularOptions={translatedSchedularOptions}
-          status={isSaveLoading ? 'waiting' : 'initial'}
-          onBack={onDialogCancel}
-          onDeploy={onDialogSave}
+          value={scheduleValue}
+          onChange={setScheduleValue}
+          onValidityChange={setIsScheduleValid}
         />
+        <div className="d-flex justify-end gap-2 m-t-md">
+          <Button
+            data-testid="back-button"
+            type="link"
+            onClick={onDialogCancel}>
+            {t('label.cancel')}
+          </Button>
+          <Button
+            data-testid="deploy-button"
+            disabled={!isScheduleValid}
+            loading={isSaveLoading}
+            type="primary"
+            onClick={onDialogSave}>
+            {t('label.save')}
+          </Button>
+        </div>
       </Modal>
     </>
   );
