--- openmetadata-ui/src/main/resources/ui/src/components/Settings/Services/AddIngestion/Steps/schedule-interval-v1.less
@@ -1,45 +1,0 @@
-/*
- *  Copyright 2025 Collate.
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *  http://www.apache.org/licenses/LICENSE-2.0
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-@import (reference) '../../../../../styles/variables.less';
-
-.schedule-interval-v1 {
-  .schedule-interval-v1-fields {
-    display: flex;
-    flex-direction: column;
-    gap: @margin-md;
-  }
-
-  .frequency-button-group {
-    display: flex;
-    flex-wrap: wrap;
-    gap: @margin-sm;
-  }
-
-  label {
-    font-weight: 500;
-  }
-
-  .cron-expression-card {
-    display: flex;
-    align-items: center;
-    gap: @margin-sm;
-    padding: @padding-sm @padding-md;
-
-    .cron-expression-card-icon {
-      height: 20px;
-      width: 20px;
-      flex-shrink: 0;
-      color: @grey-600;
-    }
-  }
-}
