--- openmetadata-ui/src/main/resources/ui/src/components/Settings/Services/AddIngestion/Steps/schedule-interval.less
@@ -1,59 +1,0 @@
-/*
- *  Copyright 2024 Collate.
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *  http://www.apache.org/licenses/LICENSE-2.0
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-@import (reference) '../../../../../styles/variables.less';
-
-.schedular-card-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 16px;
-  height: 100%;
-  .schedular-card {
-    border-radius: 6px;
-    cursor: pointer;
-    height: 100%;
-    flex: 0.5;
-    .ant-card-body {
-      padding: 24px;
-      height: 100%;
-
-      .ant-radio {
-        margin-right: 12px;
-      }
-    }
-  }
-  .schedular-card.active {
-    border-color: @primary-color;
-  }
-}
-
-.schedule-interval {
-  .ant-form-item {
-    margin: 0px;
-  }
-}
-.ant-radio-button-wrapper.week-selector-buttons,
-.ant-radio-button-wrapper.week-selector-buttons:first-child,
-.ant-radio-button-wrapper.week-selector-buttons:last-child {
-  height: 32px;
-  width: 32px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 16px;
-  border-width: 1px;
-}
-.ant-radio-button-wrapper.week-selector-buttons:not(:first-child)::before {
-  position: relative;
-}
